requirements.txt
GROQ_API_KEY = os.environ.get("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")
fastapi
uvicorn
python-multipart
pdfplumber
tesseract
poppler
